Move user search logic to user service gateway
