+++ src/shared/gateways/user-service/routes.ts
@@ -1,0 +1,5 @@
+export const USER_ROUTES = {
+    GET_USER(username: string) {
+        return `${process.env.USER_API}/users/${encodeURIComponent(username)}`;
+    },
+};
